feat: 更新脚本

+++ install-origin.sh
@@ -112,15 +112,17 @@ OS="$(uname)"
 if [[ "${OS}" == "Linux" ]]
 then
   HOMEBREW_ON_LINUX=1
-elif [[ "${OS}" != "Darwin" ]]
+elif [[ "${OS}" == "Darwin" ]]
 then
+  HOMEBREW_ON_MACOS=1
+else
   abort "Homebrew is only supported on macOS and Linux."
 fi
 
 # Required installation paths. To install elsewhere (which is unsupported)
 # you can untar https://github.com/Homebrew/brew/tarball/master
 # anywhere you like.
-if [[ -z "${HOMEBREW_ON_LINUX-}" ]]
+if [[ -n "${HOMEBREW_ON_MACOS-}" ]]
 then
   UNAME_MACHINE="$(/usr/bin/uname -m)"
 
@@ -146,9 +148,9 @@ then
 else
   UNAME_MACHINE="$(uname -m)"
 
-  # On Linux, it installs to /home/linuxbrew/.linuxbrew if you have sudo access
-  # and ~/.linuxbrew (which is unsupported) if run interactively.
-  HOMEBREW_PREFIX_DEFAULT="/home/linuxbrew/.linuxbrew"
+  # On Linux, this script installs to /home/linuxbrew/.linuxbrew only
+  HOMEBREW_PREFIX="/home/linuxbrew/.linuxbrew"
+  HOMEBREW_REPOSITORY="${HOMEBREW_PREFIX}/Homebrew"
   HOMEBREW_CACHE="${HOME}/.cache/Homebrew"
 
   STAT_PRINTF=("stat" "--printf")
@@ -179,9 +181,9 @@ fi
 export HOMEBREW_{BREW,CORE}_GIT_REMOTE
 
 # TODO: bump version when new macOS is released or announced
-MACOS_NEWEST_UNSUPPORTED="13.0"
+MACOS_NEWEST_UNSUPPORTED="14.0"
 # TODO: bump version when new macOS is released
-MACOS_OLDEST_SUPPORTED="10.15"
+MACOS_OLDEST_SUPPORTED="11.0"
 
 # For Homebrew on Linux
 REQUIRED_RUBY_VERSION=2.6    # https://github.com/Homebrew/brew/pull/6556
@@ -221,7 +223,7 @@ have_sudo_access() {
     HAVE_SUDO_ACCESS="$?"
   fi
 
-  if [[ -z "${HOMEBREW_ON_LINUX-}" ]] && [[ "${HAVE_SUDO_ACCESS}" -ne 0 ]]
+  if [[ -n "${HOMEBREW_ON_MACOS-}" ]] && [[ "${HAVE_SUDO_ACCESS}" -ne 0 ]]
   then
     abort "Need sudo access on macOS (e.g. the user ${USER} needs to be an Administrator)!"
   fi
@@ -507,41 +509,25 @@ fi
 # shellcheck disable=SC2016
 ohai 'Checking for `sudo` access (which may request your password)...'
 
-if [[ -z "${HOMEBREW_ON_LINUX-}" ]]
+if [[ -n "${HOMEBREW_ON_MACOS-}" ]]
 then
   have_sudo_access
-else
-  if [[ -w "${HOMEBREW_PREFIX_DEFAULT}" ]] ||
-     [[ -w "/home/linuxbrew" ]] ||
-     [[ -w "/home" ]]
-  then
-    HOMEBREW_PREFIX="${HOMEBREW_PREFIX_DEFAULT}"
-  elif [[ -n "${NONINTERACTIVE-}" ]]
-  then
-    if have_sudo_access
-    then
-      HOMEBREW_PREFIX="${HOMEBREW_PREFIX_DEFAULT}"
-    else
-      abort "Insufficient permissions to install Homebrew to \"${HOMEBREW_PREFIX_DEFAULT}\"."
-    fi
-  else
-    trap exit SIGINT
-    if ! /usr/bin/sudo -n -v &>/dev/null
-    then
-      ohai "Select a Homebrew installation directory:"
-      echo "- ${tty_bold}Enter your password${tty_reset} to install to ${tty_underline}${HOMEBREW_PREFIX_DEFAULT}${tty_reset} (${tty_bold}recommended${tty_reset})"
-      echo "- ${tty_bold}Press Control-D${tty_reset} to install to ${tty_underline}${HOME}/.linuxbrew${tty_reset}"
-      echo "- ${tty_bold}Press Control-C${tty_reset} to cancel installation"
-    fi
-    if have_sudo_access
-    then
-      HOMEBREW_PREFIX="${HOMEBREW_PREFIX_DEFAULT}"
-    else
-      HOMEBREW_PREFIX="${HOME}/.linuxbrew"
-    fi
-    trap - SIGINT
-  fi
-  HOMEBREW_REPOSITORY="${HOMEBREW_PREFIX}/Homebrew"
+elif ! [[ -w "${HOMEBREW_PREFIX}" ]] &&
+     ! [[ -w "/home/linuxbrew" ]] &&
+     ! [[ -w "/home" ]] &&
+     ! have_sudo_access
+then
+  abort "$(
+    cat <<EOABORT
+Insufficient permissions to install Homebrew to \"${HOMEBREW_PREFIX}\" (the default prefix).
+
+Alternative (unsupported) installation methods are available at:
+https://docs.brew.sh/Installation#alternative-installs
+
+Please note this will require most formula to build from source, a buggy, slow and energy-inefficient experience.
+We will close any issues without response for these unsupported configurations.
+EOABORT
+  )"
 fi
 HOMEBREW_CORE="${HOMEBREW_REPOSITORY}/Library/Taps/homebrew/homebrew-core"
 
@@ -559,7 +545,7 @@ EOABORT
   )"
 fi
 
-if [[ -z "${HOMEBREW_ON_LINUX-}" ]]
+if [[ -n "${HOMEBREW_ON_MACOS-}" ]]
 then
   # On macOS, support 64-bit Intel and ARM
   if [[ "${UNAME_MACHINE}" != "arm64" ]] && [[ "${UNAME_MACHINE}" != "x86_64" ]]
@@ -573,8 +559,7 @@ else
     abort "$(
       cat <<EOABORT
 Homebrew on Linux is not supported on ARM processors.
-You can try an alternate installation method instead:
-  ${tty_underline}https://docs.brew.sh/Homebrew-on-Linux#arm${tty_reset}
+  ${tty_underline}https://docs.brew.sh/Homebrew-on-Linux#arm-unsupported${tty_reset}
 EOABORT
     )"
   elif [[ "${UNAME_MACHINE}" != "x86_64" ]]
@@ -583,7 +568,7 @@ EOABORT
   fi
 fi
 
-if [[ -z "${HOMEBREW_ON_LINUX-}" ]]
+if [[ -n "${HOMEBREW_ON_MACOS-}" ]]
 then
   macos_version="$(major_minor "$(/usr/bin/sw_vers -productVersion)")"
   if version_lt "${macos_version}" "10.7"
@@ -769,10 +754,10 @@ then
   additional_shellenv_commands+=("export HOMEBREW_CORE_GIT_REMOTE=\"${HOMEBREW_CORE_GIT_REMOTE}\"")
 fi
 
-if [[ -n "${HOMEBREW_INSTALL_FROM_API-}" ]]
+if [[ -n "${HOMEBREW_NO_INSTALL_FROM_API-}" ]]
 then
-  ohai "HOMEBREW_INSTALL_FROM_API is set."
-  echo "Homebrew/homebrew-core will not be tapped during this ${tty_bold}install${tty_reset} run."
+  ohai "HOMEBREW_NO_INSTALL_FROM_API is set."
+  echo "Homebrew/homebrew-core will be tapped during this ${tty_bold}install${tty_reset} run."
 fi
 
 if [[ -z "${NONINTERACTIVE-}" ]]
@@ -827,7 +812,7 @@ execute_sudo "${CHOWN[@]}" "-R" "${USER}:${GROUP}" "${HOMEBREW_REPOSITORY}"
 
 if ! [[ -d "${HOMEBREW_CACHE}" ]]
 then
-  if [[ -z "${HOMEBREW_ON_LINUX-}" ]]
+  if [[ -n "${HOMEBREW_ON_MACOS-}" ]]
   then
     execute_sudo "${MKDIR[@]}" "${HOMEBREW_CACHE}"
   else
@@ -879,13 +864,13 @@ fi
 if should_install_command_line_tools && test -t 0
 then
   ohai "Installing the Command Line Tools (expect a GUI popup):"
-  execute_sudo "/usr/bin/xcode-select" "--install"
+  execute "/usr/bin/xcode-select" "--install"
   echo "Press any key when the installation has completed."
   getc
   execute_sudo "/usr/bin/xcode-select" "--switch" "/Library/Developer/CommandLineTools"
 fi
 
-if [[ -z "${HOMEBREW_ON_LINUX-}" ]] && ! output="$(/usr/bin/xcrun clang 2>&1)" && [[ "${output}" == *"license"* ]]
+if [[ -n "${HOMEBREW_ON_MACOS-}" ]] && ! output="$(/usr/bin/xcrun clang 2>&1)" && [[ "${output}" == *"license"* ]]
 then
   abort "$(
     cat <<EOABORT
@@ -902,14 +887,17 @@ ohai "Downloading and installing Homebrew..."
   cd "${HOMEBREW_REPOSITORY}" >/dev/null || return
 
   # we do it in four steps to avoid merge errors when reinstalling
-  execute "git" "init" "-q"
+  execute "git" "-c" "init.defaultBranch=master" "init" "--quiet"
 
   # "git remote add" will fail if the remote is defined in the global config
   execute "git" "config" "remote.origin.url" "${HOMEBREW_BREW_GIT_REMOTE}"
   execute "git" "config" "remote.origin.fetch" "+refs/heads/*:refs/remotes/origin/*"
 
   # ensure we don't munge line endings on checkout
-  execute "git" "config" "core.autocrlf" "false"
+  execute "git" "config" "--bool" "core.autocrlf" "false"
+
+  # make sure symlinks are saved as-is
+  execute "git" "config" "--bool" "core.symlinks" "true"
 
   execute "git" "fetch" "--force" "origin"
   execute "git" "fetch" "--force" "--tags" "origin"
@@ -926,26 +914,20 @@ ohai "Downloading and installing Homebrew..."
     fi
   fi
 
-  if [[ -n "${HOMEBREW_INSTALL_FROM_API-}" ]]
+  if [[ -n "${HOMEBREW_NO_INSTALL_FROM_API-}" && ! -d "${HOMEBREW_CORE}" ]]
   then
+    # Always use single-quoted strings with `exp` expressions
     # shellcheck disable=SC2016
-    ohai 'Skip tapping homebrew/core because `$HOMEBREW_INSTALL_FROM_API` is set.'
-    # Unset HOMEBREW_DEVELOPER since it is no longer needed and causes warnings during brew update below
-    if [[ -n "${HOMEBREW_ON_LINUX-}" && (-n "${HOMEBREW_CURL_PATH-}" || -n "${HOMEBREW_GIT_PATH-}") ]]
-    then
-      export -n HOMEBREW_DEVELOPER
-    fi
-  elif [[ ! -d "${HOMEBREW_CORE}" ]]
-  then
-    ohai "Tapping homebrew/core"
+    ohai 'Tapping homebrew/core because `$HOMEBREW_NO_INSTALL_FROM_API` is set.'
     (
       execute "${MKDIR[@]}" "${HOMEBREW_CORE}"
       cd "${HOMEBREW_CORE}" >/dev/null || return
 
-      execute "git" "init" "-q"
+      execute "git" "-c" "init.defaultBranch=master" "init" "--quiet"
       execute "git" "config" "remote.origin.url" "${HOMEBREW_CORE_GIT_REMOTE}"
       execute "git" "config" "remote.origin.fetch" "+refs/heads/*:refs/remotes/origin/*"
-      execute "git" "config" "core.autocrlf" "false"
+      execute "git" "config" "--bool" "core.autocrlf" "false"
+      execute "git" "config" "--bool" "core.symlinks" "true"
       execute "git" "fetch" "--force" "origin" "refs/heads/master:refs/remotes/origin/master"
       execute "git" "remote" "set-head" "origin" "--auto" >/dev/null
       execute "git" "reset" "--hard" "origin/master"
@@ -1012,16 +994,11 @@ case "${SHELL}" in
     ;;
 esac
 
-# `which` is a shell function defined above.
-# shellcheck disable=SC2230
-if [[ "$(which brew)" != "${HOMEBREW_PREFIX}/bin/brew" ]]
-then
-  cat <<EOS
+cat <<EOS
 - Run these two commands in your terminal to add Homebrew to your ${tty_bold}PATH${tty_reset}:
-    echo 'eval "\$(${HOMEBREW_PREFIX}/bin/brew shellenv)"' >> ${shell_profile}
+    (echo; echo 'eval "\$(${HOMEBREW_PREFIX}/bin/brew shellenv)"') >> ${shell_profile}
     eval "\$(${HOMEBREW_PREFIX}/bin/brew shellenv)"
 EOS
-fi
 if [[ -n "${non_default_repos}" ]]
 then
   plural=""
@@ -1029,7 +1006,8 @@ then
   then
     plural="s"
   fi
-  echo "- Run these commands in your terminal to add the non-default Git remote${plural} for ${non_default_repos}:"
+  printf -- "- Run these commands in your terminal to add the non-default Git remote%s for %s:\n" "${plural}" "${non_default_repos}"
+  printf "    echo '# Set PATH, MANPATH, etc., for Homebrew.' >> %s\n" "${shell_profile}"
   printf "    echo '%s' >> ${shell_profile}\n" "${additional_shellenv_commands[@]}"
   printf "    %s\n" "${additional_shellenv_commands[@]}"
 fi
